feat: Add WebAuthn support for login functionality
Add WebAuthn support to the login functionality by introducing new files and dependencies. This includes adding a new JSON file for storing credentials, TypeScript declarations for WebAuthn functions, and server-side and client-side code for registration and authentication. Additionally, update the package.json file to include the "@passwordless-id/webauthn" dependency.__ there are some chnaged buttons, its not permanant it was just easier to play around with webauthn that way, ill put it back to how the figma looks once i completly figure it out