feat: Add IconSvgProps type for SVG icon components
Add a new type `IconSvgProps` to the `types.ts` file in the `Components/data` directory. This type extends `SVGProps<SVGSVGElement>` and includes an optional `size` property. The `IconSvgProps` type will be used for SVG icon components throughout the application, providing a standardized interface for passing props and allowing for easy customization of icon sizes.