chore: Update deployment workflow names and file paths and docker container details
Update the names and file paths in the deployment workflows to reflect the changes made in the codebase. This includes renaming the workflow for deploying to the live site from "Deploy Web site to Live site" to "Deploy Prod Dashboard" and updating the file path for the docker-compose file in the development workflow from "frontend/occupi-web-dev/docker-compose.yml" to "frontend/occupi-web-dev/docker-compose.dev.yml". These changes ensure that the deployment workflows accurately reflect the current state of the codebase and improve clarity for future development and deployment processes.